fix(#343): move dev harness out of tests/ collection (fixes CI hang)
CI Fast Tests timed out (10 min) because pytest.ini sets testpaths=tests,
so `pytest -m "smoke or unit"` imports every module under tests/ during
collection — including the new measure_cover_grid harness. The prior
commit added a module-level sys.exit() (fires when CWN_TEST_PW is unset,
as in CI), which raised SystemExit on an xdist worker during collection,
killed the worker, and hung the run to the timeout. The first commit
passed only because its harness module imported cleanly and had no
test_ functions.

Move the harness to scripts/manual/measure_cover_grid.py (outside
testpaths; scripts/dev is gitignored so scripts/manual is used) and make
it fully import-safe: the playwright import, credential read, sys.exit,
and run all live inside main() under `if __name__ == "__main__"`.
Verified `pytest -m "smoke or unit" --collect-only` collects 1313 items
with 0 errors and never imports the harness. CSS comment + test docstring
updated to the new path. The CSS fix itself is unchanged.
